Move Create Dataset Event button above tabs
